Install prebuilt bundles from an async job, remove comments
The slow instance command now only flips executionMode and enqueues one
InstallPrebuiltLogicFunctionBundlesJob per workspace, so a slow install
never blocks the upgrade. The job installs in batches of 20 and tolerates
per-function failures; the executor still installs on-demand when a node
has no bundle.

Also fixes the RETURNING result handling: TypeORM returns [rows, count]
for UPDATE ... RETURNING, so the rows were read from the wrapper and the
enqueued job carried null ids.
